+++ docs/changes.rst
@@ -10,7 +10,11 @@ This change log is used to track all major changes to Mopidy.
 - Merged the ``gstreamer`` branch from Thomas Adamcik, bringing more than a
   hundred new tests, several bugfixes and a new backend for playing music from
   a local music archive.
-- **[WIP]** Changed backend API for get/filter/find_exact/search.
+- Backend API changes:
+
+    - Removed ``backend.playback.volume`` wrapper. Use ``backend.mixer.volume``
+      directly.
+    - **[WIP]** Changed API for get/filter/find_exact/search.
 
 
 0.1.0a0 (2010-03-27)